+++ package.json
@@ -33,8 +33,8 @@
     "@tanstack/react-query-persist-client": "^4.0.10",
     "@tanstack/react-table": "^8.5.13",
     "@thirdweb-dev/chain-icons": "^1.0.5",
-    "@thirdweb-dev/react": "^3.2.1-nightly-f669d3e",
-    "@thirdweb-dev/sdk": "^3.2.2-nightly-57b9c67",
+    "@thirdweb-dev/react": "^3.2.1",
+    "@thirdweb-dev/sdk": "^3.2.1",
     "@thirdweb-dev/storage": "^1.0.4",
     "bn.js": "^5.2.0",
     "color": "^4.0.1",
