+++ tw-components/nft-media.tsx
@@ -1,0 +1,4 @@
+import { chakra } from "@chakra-ui/react";
+import { ThirdwebNftMedia } from "@thirdweb-dev/react";
+
+export const NFTMedia = chakra(ThirdwebNftMedia);
