+++ utils/usedapp-external.ts
@@ -1,26 +1,25 @@
-import { utils } from "ethers";
+import {
+  isPossibleEVMAddress,
+  isPossibleSolanaAddress,
+} from "lib/address-utils";
 
 export function shortenString(str: string, extraShort?: true) {
   return `${str.substring(0, extraShort ? 4 : 6)}...${str.substring(
     str.length - (extraShort ? 3 : 4),
   )}`;
 }
-
-export function shortenAddress(address: string, extraShort?: true): string {
-  try {
-    const formattedAddress = utils.getAddress(address);
-    return shortenString(formattedAddress, extraShort);
-  } catch {
-    return shortenString(address, extraShort);
-  }
-}
-
 export function shortenIfAddress(
   address?: string | null | false,
   extraShort?: true,
 ): string {
-  if (typeof address === "string" && address.length > 0) {
-    return shortenAddress(address, extraShort);
+  if (!address) {
+    return "";
   }
-  return address || "";
+  if (
+    typeof address === "string" &&
+    (isPossibleSolanaAddress(address) || isPossibleEVMAddress(address, true))
+  ) {
+    return shortenString(address, extraShort);
+  }
+  return address ? address : "";
 }
